+++ CHANGELOG.rst
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 
 The format is based on `Keep a Changelog <http://keepachangelog.com/>`_.
 ====================
- Unreleased
+ 5.3.6 - 2022-12-12
 ====================
 
 Fixed